medran_course_system/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2
